move treasure chest model to client.models package
